feat: 작성자만 게시물을 삭제할 수 있도록 구현

+++ crewtalk/src/main/java/com/hh99_crewtalk/crewtalk/service/ArticleService.java
@@ -82,7 +82,13 @@ public class ArticleService {
 
     //게시물 삭제
     @Transactional
-    public Long deleteArticle(Long id) {
+    public Long deleteArticle(Long id, String username) {
+        Article article = articleRepository.findById(id).orElseThrow(() -> new InvalidArticleIdException());
+
+        if (article.getMember().getUsername() != username) {
+            throw new NotAuthorizedException();
+        }
+
         articleRepository.deleteById(id);
         return id;
     }
